refactor(racing-car): refactor to get rid of sequesce()
